+++ src/http.cpp
@@ -7,6 +7,10 @@
 #include <shared_mutex>
 
 namespace Emilia::Http {
+	std::vector<std::string> const Server::storyworlds = {
+		"erlija-past",
+		"reflections-on-blackfeather"};
+
 	Worker::Worker(
 		NativeSocket nativeSocket,
 		SocketInterface *interrupter,
@@ -36,95 +40,46 @@ namespace Emilia::Http {
 		return req;
 	}
 	std::vector<Worker::RequestFilter> Worker::filters() {
-		return {// Server status is a theme-agnostic endpoint.
-						// gilgamesh.cc, localhost, 127.0.0.1, or ::1. Refer
-						// to <https://en.cppreference.com/w/cpp/regex/ecmascript>.
-						{"(?:gilgamesh.cc|localhost|127\\.0\\.0\\.1|::1)(?::.*)?",
-						 "/status/?(?:\\?[^#]*)?(?:#.*)?",
-						 {Method::GET},
-						 &Worker::requestStatus},
-						{"(?:gilgamesh.cc|localhost|127\\.0\\.0\\.1|::1)(?::.*)?",
-						 "(/[^\\?#]*)(\\?[^#]*)?(?:#.*)?",
-						 {Method::GET},
-						 &Worker::request}};
-	}
-	Worker::ResponseAction Worker::request(
-		Request &req,
-		std::smatch const &match) {
-		// Theme is either specified as a cookie, query parameter, or given a
-		// server-side default.
-		Rain::Networking::Http::QueryParams queryParams(match[2]);
-		std::string theme{queryParams["theme"]};
-		if (theme.empty()) {
-			theme = req.headers.cookie()["theme"];
-		}
-		if (theme.empty()) {
-			// Server-side default theme.
-			theme = "erlija-past";
-		}
+		// Hosts are gilgamesh.cc, localhost, 127.0.0.1, or ::1. Refer
+		// to <https://en.cppreference.com/w/cpp/regex/ecmascript>.
+		static std::string const hostRegex{
+			"(?:gilgamesh.cc|localhost|127\\.0\\.0\\.1|::1)(?::.*)?"},
+			queryFragment{"(?:\\?[^#]*)?(?:#.*)?"};
 
-		// Request the theme-associated static.
-		std::optional<std::filesystem::path> file =
-			this->resolveThemedPath(match[1], theme);
-		if (!file) {
-			return {{StatusCode::NOT_FOUND}};
+		std::string storyworldsOr = this->server.storyworlds[0];
+		for (auto const &storyworld : this->server.storyworlds) {
+			storyworldsOr += "|" + storyworld;
 		}
-		return this->requestStatic(file.value());
-	}
-	std::optional<std::filesystem::path> Worker::resolveThemedPath(
-		std::string const &path,
-		std::string const &theme) {
-		static std::string const staticRoot{"../static"};
-
-		// path begins with / and may or may not end with a trailing /.
-		//
-		// First attempt to resolve the path at ../static/theme/{path}, then
-		// ../static/{path}.
-		static auto const resolvePath =
-			[](std::string const &pathStr) -> std::optional<std::filesystem::path> {
-			std::filesystem::path path{pathStr};
-			// If neither file nor directory exists, resolve with `.html` appended.
-			if (!std::filesystem::exists(path)) {
-				path += ".html";
-			}
-			if (!std::filesystem::exists(path)) {
-				return {};
-			}
-			path = std::filesystem::canonical(path);
-			if (std::filesystem::is_directory(path)) {
-				path /= "index.html";
-			}
-			// All files under "../static" are fair game.
-			if (!Rain::Filesystem::isSubpath(path, staticRoot)) {
-				return {};
-			}
-			return {path};
-		};
-
-		std::optional<std::filesystem::path> result =
-			resolvePath(staticRoot + "/" + theme + path);
-		if (result) {
-			return result;
-		}
-		return resolvePath(staticRoot + path);
-	}
-	Worker::ResponseAction Worker::requestStatic(
-		std::filesystem::path const &path) {
-		std::ifstream file(path, std::ios::binary);
-		std::string bytes;
-		bytes.assign(
-			(std::istreambuf_iterator<char>(file)),
-			(std::istreambuf_iterator<char>()));
-
 		return {
-			{StatusCode::OK,
-			 {{{"Content-Type", MediaType(path.extension().string())},
-				 // ~17 minute cache.
-				 {"Cache-Control", "max-age=1024"},
-				 {"Access-Control-Allow-Origin", "*"}}},
-			 std::move(bytes)}};
+			// Service status.
+			{hostRegex,
+			 "/api/status/?" + queryFragment,
+			 {Method::GET},
+			 &Worker::requestApiStatus},
+			// Themed storyworld defaults.
+			{hostRegex,
+			 "/api/storyworlds/defaults/(light|dark)",
+			 {Method::GET},
+			 &Worker::requestApiStoryworldsDefaults},
+			// General-agnostic and dependent public endpoints, which request
+			// index.html and may have a storyworld preference.
+			{hostRegex,
+			 "/(storyworlds|snapshots/(?:[^\\?#\\.]+))?" + queryFragment,
+			 {Method::GET},
+			 &Worker::requestDependentOrResolution},
+			// Force storyworld sessions. Must be updated with any new storyworld IDs.
+			{hostRegex,
+			 "/(?:" + storyworldsOr +
+				 ")(?:/(?:storyworlds|snapshots/(?:[^\\?#\\.]+))?)?" + queryFragment,
+			 {Method::GET},
+			 &Worker::requestForcedDependent},
+			// Endpoints for other statics without storyworld preference.
+			{hostRegex,
+			 "/([^\\?#]*)" + queryFragment,
+			 {Method::GET},
+			 &Worker::requestSpecificOrStaticAgnostic}};
 	}
-	Worker::ResponseAction Worker::requestStatus(
+	Worker::ResponseAction Worker::requestApiStatus(
 		Request &req,
 		std::smatch const &) {
 		using namespace Rain::Literal;
@@ -202,6 +157,146 @@ namespace Emilia::Http {
 			 {{{"Content-Type", "text/plain"},
 				 {"Content-Length", std::to_string(ssLen)}}},
 			 std::move(*ss.rdbuf())}};
+	}
+	Worker::ResponseAction Worker::requestApiStoryworldsDefaults(
+		Request &,
+		std::smatch const &match) {
+		std::string storyworld{
+			match[1] == "light" ? "erlija-past" : "reflections-on-blackfeather"};
+		return {
+			{StatusCode::OK,
+			 {{{"Content-Type", "text/plain"},
+				 {"Content-Length", std::to_string(storyworld.length())}}},
+			 storyworld}};
+	}
+	Worker::ResponseAction Worker::requestDependentOrResolution(
+		Request &req,
+		std::smatch const &match) {
+		// If a storyworld cookie is set, respond with the index page for that
+		// storyworld. Then, check if the path has a preferred storyworld.
+		// Otherwise, return with the catch-all index.
+		auto storyworld = this->resolveStoryworld(req, match[1]);
+		auto file = this->resolveStoryworldPath("/index.html", storyworld.first);
+		if (!file) {
+			// Should not occur as long as index.html exists for all storyworlds and
+			// catch-all.
+			return {{StatusCode::NOT_FOUND}};
+		}
+		auto response = this->requestStatic(file.value());
+		if (storyworld.second) {
+			// Set preferred storyworld if determined that way. The cookie lasts for
+			// the entire session.
+			response.response->headers.setCookie(
+				{{"storyworld-preferred",
+					{storyworld.first.value(), {{"Path", "/"}}}}});
+		}
+		return response;
+	}
+	std::pair<std::optional<std::string>, bool> Worker::resolveStoryworld(
+		Request &req,
+		std::string const &path) {
+		auto cookie = req.headers.cookie();
+		auto it = cookie.find("storyworld-selected");
+		if (it != cookie.end()) {
+			return {it->second, false};
+		}
+		static std::unordered_map<std::string, std::string> const preferences{
+			{"snapshots/see-it-wasnt-always-like-this", "erlija-past"}};
+		auto preference = preferences.find(path);
+		if (preference != preferences.end()) {
+			return {preference->second, true};
+		}
+		it = cookie.find("storyworld-preferred");
+		if (it != cookie.end()) {
+			return {it->second, false};
+		}
+		it = cookie.find("storyworld-defaulted");
+		if (it != cookie.end()) {
+			return {it->second, false};
+		}
+		return {{}, false};
+	}
+	std::optional<std::filesystem::path> Worker::resolveStoryworldPath(
+		std::string const &path,
+		std::optional<std::string> const &storyworld) {
+		static std::string const staticRoot{"../static"};
+
+		// path does not begin with / and may or may not end with a trailing /.
+		//
+		// First attempt to resolve the path at ../static/{storyworld}/{path}, then
+		// ../static/{path}.
+		static auto const resolvePath =
+			[](std::string const &pathStr) -> std::optional<std::filesystem::path> {
+			std::filesystem::path path{pathStr};
+			if (!std::filesystem::exists(path)) {
+				return {};
+			}
+			path = std::filesystem::canonical(path);
+			// All files under "../static" are fair game.
+			if (!Rain::Filesystem::isSubpath(path, staticRoot)) {
+				return {};
+			}
+			return {path};
+		};
+		if (storyworld) {
+			std::optional<std::filesystem::path> result =
+				resolvePath(staticRoot + "/" + storyworld.value() + "/" + path);
+			if (result) {
+				return result;
+			}
+		}
+		return resolvePath(staticRoot + "/" + path);
+	}
+	Worker::ResponseAction Worker::requestStatic(
+		std::filesystem::path const &path) {
+		std::ifstream file(path, std::ios::binary);
+		std::string bytes;
+		bytes.assign(
+			(std::istreambuf_iterator<char>(file)),
+			(std::istreambuf_iterator<char>()));
+
+		return {
+			{StatusCode::OK,
+			 {{{"Content-Type", MediaType(path.extension().string())},
+				 // No cache by default.
+				 {"Cache-Control", "no-store"},
+				 {"Access-Control-Allow-Origin", "*"}}},
+			 std::move(bytes)}};
+	}
+	Worker::ResponseAction Worker::requestForcedDependent(
+		Request &,
+		std::smatch const &) {
+		auto file = this->resolveStoryworldPath("/forced.html", {});
+		if (!file) {
+			return {{StatusCode::NOT_FOUND}};
+		}
+		return this->requestStatic(file.value());
+	}
+	Worker::ResponseAction Worker::requestSpecificOrStaticAgnostic(
+		Request &req,
+		std::smatch const &match) {
+		auto storyworld = this->resolveStoryworld(req, match[1]);
+		auto file = this->resolveStoryworldPath(match[1], storyworld.first);
+		if (!file) {
+			return {{StatusCode::NOT_FOUND}};
+		}
+		auto response = this->requestStatic(file.value());
+
+		// Only select files are cached.
+		// TODO: Implement a more robust policy.
+		static std::set<std::string> const noCache{
+			"index.html",
+			"index.css",
+			"index.js",
+			"colors.css",
+			"fonts.css",
+			"components/snapshot.html",
+			"components/snapshot.css",
+			"components/snapshot.js"};
+		if (noCache.find(match[1]) == noCache.end()) {
+			response.response.value().headers["Cache-Control"] = "Max-Age=3600";
+		}
+		return response;
 	}
 
 	Server::Server(